+++ src/Task/Process/ProcessLauncherTask.php
@@ -220,7 +220,7 @@ class ProcessLauncherTask extends AbstractConfigurableTask implements FlushableT
         $resolver->setAllowedTypes('process_options', ['array']);
         $resolver->setNormalizer(
             'process_options',
-            static function (Options $options, $value): int|float|string|bool|null {
+            static function (Options $options, $value): array {
                 if (!empty($value)) {
                     // Todo deprecation trigger
                     throw new \InvalidArgumentException('Deprecated option, please contact support for help');